backend/requirements.txt
# Procrastify Backend Dependencies
flask>=3.0.0
flask-cors>=4.0.0
mysql-connector-python>=8.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
